changed "Stars" float -> byte

+++ src/PropertyManager/Travely.PropertyManager.Data/Models/Property.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
 
         public string Name { get; set; }
-        public float Stars { get; set; }
+        public byte Stars { get; set; }
         public string Address { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
